+++ widgets/views/push-notification-info.php
@@ -1,7 +1,0 @@
-<?php
-
-use humhub\widgets\bootstrap\Alert;
-
-?>
-
-<?= Alert::danger('<strong>Push Notifications</strong> not allowed.') ?>