+++ docs/CHANGELOG.md
@@ -1,11 +1,15 @@
 Changelog
 =========
 
-2.2.0 (July 26, 2025)
----------------------
+2.2.0 (Unreleased)
+----------------------
+- Enh #74: Move non Firebase stuff to core
+- Enh #79: Migration to Bootstrap 5 for HumHub 1.18
+
+2.1.9 (Unreleased)
+----------------------
 - Fix #75: Warning for "FCM Detection - App is using Proxy Firebase Service"
 - Fix #78: Don't show opener if logout triggered by mobile app
-- Enh #79: Migration to Bootstrap 5 for HumHub 1.18
 
 2.1.8 (March 21, 2025)
 ----------------------